support for exports section
Our utf8 handling is too bloated. We are using

typedef {
 char *str
 u32 str;
} utf8_t;

to represent a UTF-8 encoded string. It looks like it is legal for
a properly UTF-8 encoded string to contain NUL, so we can't just use
C strings (though in many places, it seems people just use C strings
for UTF-8 and hope for the best).

There must be a more compact representation for UTF-8 strings in C.
Need to find one.